2.Utilizando o vertex no python/hello_world_vertex/pyproject.toml
[project]
name = "hello-world-vertex"
version = "0.1.0"
description = "A very simple python script calling vertex ai"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-aiplatform>=1.134.0",
    "google-genai>=1.60.0",
    "python-decouple>=3.8",
]
